More code reorganization.
